Add some tests for the Rails' generators
Add basic tests for the Rails generators provided with Bear. Basically
just create a new Rails application and run most of the available
generators and their options.

Rely on Aruba to check file/folder presence and content and also for
steps related to the command line.

Bear requires Rails as a _development_ dependency, it could work with
Rails 3.1+ but support for these versions is not on the roadmap.